+++ examples/src/main/kotlin/examples/onnx/cv/efficicentnet/notop/EfficientNetB0additionalTrainingWithTensorFlow.kt
@@ -70,8 +70,8 @@ fun efficientNetB0AdditionalTraining() {
         println(it)
         val preprocessing = pipeline<BufferedImage>()
             .resize {
-                outputHeight = IMAGE_SIZE.toInt()
-                outputWidth = IMAGE_SIZE.toInt()
+                outputHeight = IMAGE_SIZE
+                outputWidth = IMAGE_SIZE
                 interpolation = InterpolationType.BILINEAR
             }
             .convert { colorMode = ColorMode.BGR }
