+++ examples/src/main/kotlin/examples/transferlearning/modelhub/vgg16/Example_5_VGG16_additional_training_noTop.kt
@@ -61,8 +61,8 @@ fun vgg16noTopAdditionalTraining() {
 
     val preprocessing = pipeline<BufferedImage>()
         .resize {
-            outputHeight = IMAGE_SIZE.toInt()
-            outputWidth = IMAGE_SIZE.toInt()
+            outputHeight = IMAGE_SIZE
+            outputWidth = IMAGE_SIZE
             interpolation = InterpolationType.BILINEAR
         }
         .convert { colorMode = ColorMode.BGR }
